move mongo config from db to configs
