+++ server/src/tests/integration/message.test.ts
@@ -1,7 +1,7 @@
 import request from 'supertest';
 import app from '@/app.js';
 import mongoose from 'mongoose';
-import connectMongoDB from '@/db/mongo.js';
+import connectMongoDB from '@/configs/mongo.js';
 import Message from '@/models/message.model.js';
 import prisma from '@/configs/prismaClient.js';
 import { DateTime } from 'luxon';
